@@ -13,17 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
+    "@angular/cli": "^8.1.3",
+    "@angular/common": "^8.1.3",
+    "@angular/core": "^8.1.3",
+    "@angular/forms": "^8.1.3",
     "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
+    "@angular/platform-browser": "^8.1.3",
+    "@angular/platform-browser-dynamic": "^8.1.3",
+    "@angular/router": "^8.1.3",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^4.6.2",
+    "@ionic/angular": "^4.7.1",
     "@ionic/storage": "^2.2.0",
     "@types/echarts": "^4.1.9",
     "cordova-android": "8.0.0",
@@ -35,18 +36,17 @@
     "rxjs": "~6.5.1",
     "swiper": "^4.5.0",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.13.8",
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular-devkit/core": "~7.3.8",
-    "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
-    "@ionic/angular-toolkit": "~1.5.1",
+    "@angular-devkit/architect": "^0.801.3",
+    "@angular-devkit/build-angular": "^0.801.3",
+    "@angular-devkit/core": "^8.1.3",
+    "@angular-devkit/schematics": "^8.1.3",
+    "@angular/compiler": "~8.1.3",
+    "@angular/compiler-cli": "~8.1.3",
+    "@angular/language-service": "~8.1.3",
+    "@ionic/angular-toolkit": "^2.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~12.0.0",
@@ -68,7 +68,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~8.3.0",
     "tslint": "~5.17.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.5"
   },
   "description": "An Ionic project",
   "cordova": {
