@@ -24,9 +24,9 @@ export interface qryGameForm {
 export class GamePage implements OnInit {
 
 
-  @ViewChild(IonRefresher) refresher: IonRefresher;
-  @ViewChild(IonInfiniteScroll) infiniteScroll: IonInfiniteScroll;
-  @ViewChild(IonContent) content: IonContent;
+  @ViewChild(IonRefresher, { static: false }) refresher: IonRefresher;
+  @ViewChild(IonInfiniteScroll, { static: false }) infiniteScroll: IonInfiniteScroll;
+  @ViewChild(IonContent, { static: false }) content: IonContent;
 
 
   formData: qryGameForm = {
@@ -56,6 +56,7 @@ export class GamePage implements OnInit {
   }
 
   ngOnInit() {
+    console.log(this.refresher)
     let curDate = new Date(new Date().getTime());
     let nextDate = new Date(curDate.getTime() + 24 * 60 * 60 * 1000); //后一天
     this.beginTime = this.utils.dateFormat(curDate, 'YYYY-MM-DDTHH:mm+08:00');
