@@ -14,8 +14,8 @@ import { map, finalize } from 'rxjs/operators';
 })
 export class GamePage implements OnInit {
 
-  @ViewChild(IonRefresher) refresher: IonRefresher;
-  @ViewChild(IonContent) content: IonContent;
+  @ViewChild(IonRefresher, { static: false }) refresher: IonRefresher;
+  @ViewChild(IonContent, { static: false }) content: IonContent;
 
   formData = {
     beginTime: '2019-04-01',
