@@ -17,8 +17,8 @@ import { IonRefresher, IonContent, ModalController, AlertController } from '@ion
 export class GoldChgRankPage implements OnInit {
 
 
-  @ViewChild(IonRefresher) refresher: IonRefresher;
-  @ViewChild(IonContent) content: IonContent;
+  @ViewChild(IonRefresher, { static: false }) refresher: IonRefresher;
+  @ViewChild(IonContent, { static: false }) content: IonContent;
   formData = {
     page: 1,
     limit: 10,
