@@ -15,8 +15,8 @@ import { IonRefresher, IonContent, ModalController, NavController } from '@ionic
 export class OperatePage implements OnInit {
 
 
-  @ViewChild(IonRefresher) refresher: IonRefresher;
-  @ViewChild(IonContent) content: IonContent;
+  @ViewChild(IonRefresher, { static: false }) refresher: IonRefresher;
+  @ViewChild(IonContent, { static: false }) content: IonContent;
 
   formData = {
     beginTime: '2019-04-01',
