@@ -15,8 +15,8 @@ import { SuccessOrderPage } from './success-order/success-order.page';
 })
 export class PayPage implements OnInit {
 
-  @ViewChild(IonRefresher) refresher: IonRefresher;
-  @ViewChild(IonContent) content: IonContent;
+  @ViewChild(IonRefresher, { static: false }) refresher: IonRefresher;
+  @ViewChild(IonContent, { static: false }) content: IonContent;
 
   formData = {
     beginTime: '2019-04-01',
