@@ -4,10 +4,10 @@ import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 const routes: Routes = [
   {
     path: 'module',
-    loadChildren: './main/main.module#MainPageModule'
+    loadChildren: () => import('./main/main.module').then(mod => mod.MainPageModule)
   },
-  { path: 'login', loadChildren: './login/login.module#LoginPageModule' },
-  { path: 'page404', loadChildren: './page404/page404.module#Page404PageModule' },
+  { path: 'login', loadChildren: () => import('./login/login.module').then(mod => mod.LoginPageModule) },
+  { path: 'page404', loadChildren: () => import('./page404/page404.module').then(mod => mod.Page404PageModule) },
 
   { path: '', redirectTo: '/module/home', pathMatch: 'full' },
   { path: '**', redirectTo: '/page404' },
