@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 const routes: Routes = [
-  { path: 'schedule', loadChildren: '../job/schedule/schedule.module#SchedulePageModule' },
+  { path: 'schedule', loadChildren: () =>import('../job/schedule/schedule.module').then(mod => mod.SchedulePageModule) },
 ];
 
 @NgModule({
