@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 const routes: Routes = [
-  { path: 'oss', loadChildren: './oss/oss.module#OssPageModule' },
+  { path: 'oss', loadChildren: () => import('./oss/oss.module').then(mod => mod.OssPageModule) },
 ];
 
 @NgModule({
